Refactor `ringslide` struct to private header
